backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.0
langchain-groq==0.2.0
langgraph==0.2.28
langchain-core==0.3.5
groq==0.11.0
httpx==0.27.2
python-multipart==0.0.12
alembic==1.13.3
